Fix minor review notes and reuse shared byte-format utility in FastAPI
Agent-Logs-Url: https://github.com/MicheleDelliVeneri/ALMASim/sessions/ca01ea85-70a6-41f6-ada8-713b660e192c

+++ src/almasim/services/simulation.py
@@ -130,7 +130,7 @@ class SimulationParams:
         freq_support: str,
         cont_sens: float,
         antenna_array: str,
-        n_pix: Optional[float],
+        n_pix: Optional[int],
         n_channels: Optional[int],
         source_type: str,
         tng_api_key: Optional[str],
